Implement jump table fix-ups for linear register allocator.
We achieve this by splitting up instruction selection for case
switches into two parts: the actual code generation, and the
generation of the accompanying jump table.  With this scheme,
the jump fixup code can modify the contents of the jump table
stored within the JMP_TBL (or BCTL) instruction, before the
actual data section is created.

SPARC and PPC patches are untested; they might not work!

Signed-off-by: Edward Z. Yang <ezyang@mit.edu>