Fix bug in one-line stats printing
The code that prints the "one-line" stats (i.e. the RTS -t flag) was
incorreclty printing zeros for some time values. These time values
were computed inside a conditional that was only true when printing
detailed stats (i.e. the RTS -s or -S flags).

This commit simply moves the computation out of the conditional so
they are available for the one-line stats output.