Keep the C stack pointer 16-byte aligned on all x86 platforms, not just Mac OS X (#5250).
The OS X ABI requires the C stack pointer to be 16-byte aligned at a
function call.  As far as I know this is not a requirement on other
x86 ABIs, but it seems that gcc is now generating SSE2 code that
assumes stack alignment (-mincoming-stack-boundary defaults to 4), so
we have to respect 16-byte alignment.