Add two new primops:
  seq#   :: a -> State# s -> (# State# s, a #)
  spark# :: a -> State# s -> (# State# s, a #)

seq# is a version of seq that can be used in a State#-passing
context.  We will use it to implement Control.Exception.evaluate and
thus fix #5129.  Also we have plans to use it to fix #5262.

spark# is to seq# as par is to pseq.  That is, it creates a spark in a
State#-passing context.  We will use spark# and seq# to implement rpar
and rseq respectively in an improved implementation of the Eval monad.