defaultErrorHandler now only takes LogAction
It used to take a whole DynFlags, but that meant we had to
create a DynFlags with (panic "No settings") for settings, as
we didn't have any real settings.

Now we just pass the LogAction, which is all that it actually needed.
The default is exported from DynFlags as defaultLogAction.