Add new TrieMap module
This lets you make efficent finite mappings whose
keys are  CoreExpr
          Type
	  Coercion
	  lists
	  maybes

This is useful for CSE, and maybe elsewhere.  I've
been wanting to do this for some time, just for fun,
and it worked out quite nicely.  The code is fairly
boiler-plate-ish, so it's easy to add new key types.