Print type contexts with fsep, not sep
This is a slightly experimental change. When pretty-printing
a type, instead of
      instance (Eq a,
                Eq b,
                Eq c,
                Eq d,
                Eq e,
                Eq f,
                Eq g,
                Eq h,
                Eq i,
                Eq j,
                Eq k,
                Eq l) =>
               Eq (a, b, c, d, e, f, g, h, i, j, k, l)
you'll get

      instance (Eq a, Eq b, Eq c, Eq d, Eq e, Eq f, Eq g, Eq h, Eq i,
                Eq j, Eq k, Eq l) =>
               Eq (a, b, c, d, e, f, g, h, i, j, k, l)

That is, if the context doesn't fit on one line, we break it only
where necessary rather that after each item.

The test suite comparison ignores white space, so the change
produces no failures; its a UI thing. It's a one-character
change in TypeRep, so easily reversed.