Increase the "context stack depth" to 200 (from 20)
This parameter controls the allowed depth of reasoning in the
type constraint solver.  Perfectly well-behaved programs can
use deep stacks, and 20 is obviously too small.  (Indeed, if
you don't have UndecidableInstances, the constraint solver
is supposed to terminate, so no limit should be needed.)

Responding to Trac #5395 this patch increases the default
to 200.