Accept type error message changes
I think ideally we should give the info from both the old and new
messages; i.e. "You wrote X, which I reduced to Y, which isn't valid".
That way if Y is an abstract newtype then it will be clearer why
the type wasn't valid. We used to only give X, and now we only give Y.
Finding X is now a little messier, especially in cases like ccfail005.