Use a custom dead-code analyser instead of calling the occurrence analyser.
In CorePrep we used to call the occurrence analyser to drop dead code
(see Note [Dead code in CorePrep]), but the occurrence analyser
sometimes introduces new let bindings for case binders.  This was
leading to the bug in #5433: the let binding introduced was for an
unlifted value, and the code generator is not expecting to see let
bindings of unlifted values (CorePrep is supposed to eliminate them).
We don't want this let binding anyway, so instead of using the
occurrence analyser here we have a simple custom dead-code analyser.