Fix gratuitous code duplication in the byte code generator
For a case expression with a default, we were re-emitting the code for
the default branch each time it was referenced, which in a case with
many branches could be many times, leading to an O(n^2) blowup.

This shows up in ghcirun003, which now runs much faster than before.